+++ src/java/us/kbase/common/executionengine/LineLogger.java
@@ -1,4 +1,4 @@
-package us.kbase.mobu.tester;
+package us.kbase.common.executionengine;
 
 public interface LineLogger {
